+++ plugin_scripts/__init__.py
@@ -1,4 +1,4 @@
-from .deploy import main
+from plugin_scripts.deploy import main
 
 if __name__ == "__main__":
     main()