+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2022-09-24
+
+### Added
+
+- Option to only execute changed files
+- Option to execute DML statements
+
 ## [0.5.0] - 2022-04-11
 
 ### Updated